refactor: Update skills page with tooltip icons
The skills page has been updated to display tooltip icons instead of plain text for each skill. This improves the visual presentation and provides a more intuitive user experience. The icons used are from popular libraries such as Akar Icons, Font Awesome, Tabler, Remix Icons, and Simple Icons.

This commit refactors the code in the skills.vue file to replace the plain text with corresponding tooltip icons. Each skill now has a unique icon associated with it, making it easier for users to identify and understand the skills being showcased.

The changes include:
- Replacing the "Javascript" text with the Akar Icons JavaScript icon.
- Replacing the "Vue Js" text with the Font Awesome Vue.js icon.
- Replacing the "Nuxt Js" text with the Tabler Nuxt.js icon.
- Replacing the "Tailwind CSS" text with the Remix Icons Tailwind CSS icon.
- Replacing the "PHP" text with the Material Symbols PHP icon.
- Replacing the "Laravel" text with the Simple Icons Laravel icon.

These changes enhance the visual appeal of the skills page and provide a more engaging experience for users.